Add files via upload
Tuple Methods:
The index() method returns the index of a specified element.
The count() method returns the total occurrence of a specified element in a tuple.
zip() function takes multiple lists as arguments and zips them together.

Set Methods:
add() method adds an element to a set.
update() method helps to add multiple elements to a set.
remove() is used to remove an item/element from the set.
union() method returns the union of two sets.
intersection() method returns the intersection of two sets.
difference() method returns the difference of two sets.
symmetric_difference() method returns the set of elements in A and B but not in both (excluding the intersection).

Dictionary Methods:
keys() method returns the list of keys in the dictionary object.
values() method returns the list of values in the dictionary object.
items() method returns the list of the keys and values.
get() method takes the key as an argument and returns None if the key is not found in the dictionary.
update() method is used to add an element which is a key-value pair. We can also use the update() method to update the existing values for a key.
dict() method is used to create dictionary objects from sequence of items which are pairs.
pop() method removes and returns an element from a dictionary having the given key.