Add files via upload
A Statement is an instruction that the Python interpreter can execute.
An Expression is a combination of values, variables, operators, and calls to functions.

In Python, the end of a statement is marked by a newline character. But we can make a statement extend over multiple lines with the line continuation character ().

if Statements in Python allows us to tell the computer to perform alternative actions based on a certain set of results.