Add files via upload
for loop acts as an iterator in Python; it goes through items that are in a sequence or any other iterable item.
Another way to loop through the elements of a list is to use the enumerate() function. Using enumerate requires us two iterators index and element.
range() function is a pretty useful function to get a sequence of numbers. It takes three arguments : start, stop, step.

A while statement will repeatedly execute a single statement or group of statements as long as the condition is true.
break: Breaks out of the current closest enclosing loop.
continue: Goes to the top of the closest enclosing loop.
pass: Does nothing at all.