Add files via upload
List comprehensions are a tool for transforming one list (any iterable actually) into another list. 
During this transformation, elements can be conditionally included in the new list and each element can be transformed as needed.

List comprehension consists of the following elements:

Output Expression- input sequence- an optional conditional statement.
even_squared_numbers = [number**2 for number in list_of_numbers if number%2==0] 

Note: If there is an else statement included in the optional statement, then the optional statement must come before the input sequence.
odd_squared_numbers = [number**2 if number%2!=0 else number**3 for number in list_of_numbers ]

Nested comprehesions:
Always the outer loop comes first and inner loop is followed.(Remember the order of statements in the comprehension is the same as order in for loop)
list_of_numbers = [num2 for num1 in range(4) for num2 in range(3)]
list_of_numbers = [num2**2 for num2 in range(3) for num1 in range(4)] use these examples to understand the above statement.

List of lists:
list_of_lists = [[num2 for num2 in range(3)] for num1 in range(4)] 

For strings 
first_letters = [word[0] for word in words_by_walter] #for getting first letter of words in any list of strings.

my_string = 'Alphabet Inc.'
list_of_characters = [char for char in my_string]
print(list_of_characters)

We also have tuple comprehensions and set comprehensions. look into the file for detailed understanding.

Dictionary comprehensions:
For flipping the key and values in any dictionary items for an example with countries and cities.
flipped = { city : country for country, city in country_city_dict.items() if country[0] not in ['E','A']}